+++ src/components/html/HTML.unit.js
@@ -1,3 +1,4 @@
+import Webform from '../../Webform';
 import Harness from '../../../test/harness';
 import HTMLComponent from './HTML';
 import sinon from 'sinon';
@@ -5,7 +6,8 @@ import assert from 'power-assert';
 
 import {
   comp1,
-  comp2
+  comp2,
+  comp3,
 } from './fixtures';
 
 describe('HTML Component', () => {
@@ -29,5 +31,31 @@ describe('HTML Component', () => {
       component.checkRefreshOn(null);
       assert.equal(emit.callCount, 0);
     });
+  });
+
+  it('Should not execute scripts inside HTML component, but execute interpolation properly', (done) => {
+    const formElement = document.createElement('div');
+    const form = new Webform(formElement);
+
+    const alert = sinon.spy(window, 'alert');
+    form.setForm(comp3).then(() => {
+      setTimeout(() => {
+        assert.equal(alert.callCount, 0);
+        const div = form.element.querySelector('.myClass');
+        assert.equal(div.innerHTML.trim(), 'No Text');
+
+        const textField = form.getComponent(['textField']);
+        textField.setValue('apple', { modified: true });
+
+        setTimeout(() => {
+          const div = form.element.querySelector('.myClass');
+
+          assert.equal(div.innerHTML.trim(), 'apple');
+          assert.equal(div.className, 'myClass apple-class');
+          done();
+        }, 400);
+      }, 200);
+    })
+      .catch(done);
   });
 });