+++ src/components/select/Select.unit.js
@@ -1212,3 +1212,4 @@ describe('Select Component with Entire Object Value Property', () => {
     });
   });
 });
+